+++ README.md
@@ -70,7 +70,7 @@ graph TD
 ### 덱
 
 - [x] 카드 목록을 가진다.
-- [ ] 카드를 섞는다.
+- [x] 카드를 섞는다.
 - [x] 카드를 뽑는다.
   - [x] 카드가 없으면 뽑을 수 없다.
