+++ src/test/java/blackjack/domain/DeckTest.java
@@ -35,19 +35,3 @@ class DeckTest {
                 .isInstanceOf(IllegalStateException.class);
     }
 }
-
-class Deck {
-
-    private final Stack<Card> cards;
-
-    public Deck(final Stack<Card> cards) {
-        this.cards = cards;
-    }
-
-    public Card draw() {
-        if (cards.empty()) {
-            throw new IllegalStateException("덱에 더 이상의 카드가 없습니다.");
-        }
-        return cards.pop();
-    }
-}
